requirements.txt
Django==1.4.3
Fabric==1.5.3
South==0.7.6
Unipath==0.2.1
amqp==1.0.6
anyjson==0.3.3
argparse==1.2.1
billiard==2.7.3.19
celery==3.0.13
django-cache-panel==0.1
django-celery==3.0.11
django-debug-toolbar==0.9.4
django-extensions==1.0.3
django-pylibmc-sasl==0.2.4
gunicorn==0.17.2
httplib2==0.7.7
kombu==2.5.4
oauth2==1.5.211
paramiko==1.9.0
psycopg2==2.4.6
pycrypto==2.6
pylibmc==1.2.3
python-dateutil==1.5
python-twitter==0.8.2
pytz==2012j
simplejson==3.0.7
wsgiref==0.1.2
